+++ model/gn/@stk_model_gn_heteroscedastic0/stk_variance_eval.m
@@ -1,8 +1,15 @@
-% SUBSREF [overload base function]
+% FEVAL [overload base function]
+%
+% EXPERIMENTAL CLASS WARNING:  The stk_model_gn_heteroscedastic0 class is
+%    currently considered experimental.  STK users who wish to experiment with
+%    it are welcome to do so, but should be aware that API-breaking changes
+%    are likely to happen in future releases.  We invite them to direct any
+%    questions, remarks or comments about this experimental class to the STK
+%    mailing list.
 
 % Copyright Notice
 %
-%    Copyright (C) 2015, 2016 CentraleSupelec
+%    Copyright (C) 2018 CentraleSupelec
 %
 %    Author:  Julien Bect  <julien.bect@centralesupelec.fr>
 
@@ -26,39 +33,18 @@
 %    You should  have received a copy  of the GNU  General Public License
 %    along with STK.  If not, see <http://www.gnu.org/licenses/>.
 
-function value = subsref (M_post, idx)
+function v = stk_variance_eval (gn, x, diff)
 
-if strcmp (idx(1).type, '.')
-    
-    value = get (M_post, idx(1).subs);
-    
-    if length (idx) > 1
-        value = subsref (value, idx(2:end));
-    end
-    
-else
-    
-    stk_error ('Illegal indexing.', 'IllegalIndexing');
-    
+% Defaut value for 'diff' (arg #3): -1
+if nargin < 3,  diff = -1;  end
+
+switch diff
+    case {-1, 1}
+        % -1 means "no derivative"
+        % +1 means "derivative wrt log_dispersion"
+        v = exp (gn.log_dispersion) * feval (gn.variance_function, double (x));
+    otherwise
+        stk_error ('diff should be either -1 or +1', 'IncorrectArgument');
 end
 
 end % function
-
-
-%!shared M_post, n
-%! n = 15;
-%! x_obs = (linspace (0, pi, n))';
-%! z_obs = sin (x_obs);
-%!
-%! M_prior = stk_model ('stk_materncov32_iso');
-%! M_prior.param = log ([1.0; 2.1]);
-%!
-%! M_post = stk_model_gpposterior (M_prior, x_obs, z_obs);
-
-%!assert (isstruct (M_post.prior_model));
-%!assert (isequal (size (M_post.input_data), [n 1]));
-%!assert (isequal (size (M_post.output_data), [n 1]));
-%!assert (isa (M_post.kreq, 'stk_kreq_qr'));
-
-%!error M_post(1);
-%!error M_post{1};