+++ NEWS.md
@@ -13,7 +13,25 @@
   instead of NaN when the derivative with respect to lnv is requested
   for a noiseless model.
 
+* `stk_model_gpposterior`: Trigger parameter estimation when
+  lognoisevariance is NaN.  In the (experimental) case where
+  lognoisevariance is a noise model object, parameter estimation is
+  triggered as soon as at least one of the optimizable parameters of
+  the model is NaN.
+  ([ticket #93](https://sourceforge.net/p/kriging/tickets/93/))
+
 * Deprecated: `stk_noisecov`.
+
+## Noise model objects [experimental]
+
+* `stk_model_gn`, `stk_model_gn_heteroscedastic0`: New classes
+  representing Gaussian noise models.  Currently in experimental
+  state, these two classes are merely a proof-of-concept.  Other noise
+  model classes can be defined by subclassing `stk_model_gn`.
+  ([ticket #38](https://sourceforge.net/p/kriging/tickets/38/))
+  
+* `stk_example_misc06.m`: Demonstrate the use of noise model objects
+  to estimate the dispersion parameter in an heteroscedatic case.
 
 ## `stk_dataframe` and related classes